This version is focused on points detections. The camera stuff are still in discussion and investigation, so I prefer wait for a consensus.
So the main improvement is in PointsToTrackSIFT class and MotionProcessor class. If something is not well designed, we should talk about because as it's not directly a SfM part, it's important too.

The file testPointsToTrack.cpp can be a good entry point as it's a main test for the two classes... I'm waiting your feedbacks!